some fixes in management commands

+++ birzum/apps/products/management/commands/write_price_in_sum.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from django.core.management.base import BaseCommand
 
 from birzum.apps.smallapps.rating.models import Currency
@@ -26,12 +28,12 @@ class Command(BaseCommand):
                 # item item has price field completed
                 # complete price_sum field as well according to usd rate
                 if item.price:
-                    item.price_sum = item.price * self.usd_course.currency
+                    item.price_sum = Decimal(int(item.price * self.usd_course.currency))
 
                 # item item has price field completed
                 # complete price_sum field as well according to usd rate
                 if item.discount_price:
-                    item.discount_price_sum = item.discount_price * self.usd_course.currency
+                    item.discount_price_sum = Decimal(int(item.discount_price * self.usd_course.currency))
 
                 item.save()
